+++ tests/FilesystemTest.php
@@ -17,17 +17,17 @@ afterEach(function (): void {
     unset($this->tempDir);
 });
 
-test('test instances', function (): void {
+test('instances', function (): void {
     $this->assertInstanceOf(Filesystem::class, new Filesystem);
     $this->assertInstanceOf(File::class, new File('/1/1.txt'));
     $this->assertInstanceOf(Directory::class, new Directory('/1'));
 });
 
-test('test filesystem helper', function (): void {
+test('filesystem helper', function (): void {
     $this->assertInstanceOf(Filesystem::class, filesystem());
 });
 
-test('test deleteDirectory() method', function (): void {
+test('directory delete method', function (): void {
     @mkdir($this->tempDir . '/1');
     @mkdir($this->tempDir . '/1/2');
     @mkdir($this->tempDir . '/1/2/3');
@@ -36,35 +36,35 @@ test('test deleteDirectory() method', function (): void {
     $this->assertTrue($filesystem->directory($this->tempDir . '/1')->delete());
 });
 
-test('test replace() method', function (): void {
+test('file replace method', function (): void {
     file_put_contents($this->tempDir . '/replace.txt', 'foo');
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/replace.txt')->replace('foo', 'boo');
     $this->assertEquals('boo', file_get_contents($this->tempDir . '/replace.txt'));
 });
 
-test('test put() method', function (): void {
+test('file put method', function (): void {
     $filesystem = new Filesystem();
     $this->assertEquals(4, $filesystem->file($this->tempDir . '/2.txt')->put('test'));
 });
 
-test('test isFile() method', function (): void {
+test('file isFile method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('test');
     $this->assertTrue($filesystem->file($this->tempDir . '/1.txt')->isFile());
 });
 
-test('test isWindowsPath() method', function (): void {
+test('filesystem isWindowsPath method', function (): void {
     $filesystem = new Filesystem();
     $this->assertTrue($filesystem->isWindowsPath('C:\file\1.txt'));
 });
 
-test('test isDirectory() method', function (): void {
+test('directory isDirectory method', function (): void {
     $filesystem = new Filesystem();
     $this->assertTrue($filesystem->directory($this->tempDir)->isDirectory());
 });
 
-test('test isReadable() method', function (): void {
+test('file isReadable method', function (): void {
     if (PHP_OS_FAMILY === 'Windows') {
         $this->markTestSkipped('The operating system is Windows');
     }
@@ -81,7 +81,7 @@ test('test isReadable() method', function (): void {
 });
 
 
-test('test isWritable() method', function (): void {
+test('file isWritable method', function (): void {
 
     if (PHP_OS_FAMILY === 'Windows') {
         $this->markTestSkipped('The operating system is Windows');
@@ -103,7 +103,7 @@ test('test isStream() method', function (): void {
     $this->assertTrue($filesystem->isStream('file://1.txt'));
 });
 
-test('test isAbsolute method', function (): void {
+test('filesystem isAbsolute method', function (): void {
     $filesystem = new Filesystem();
 
     $this->assertFalse($filesystem->isAbsolute(''));
@@ -120,7 +120,7 @@ test('test isAbsolute method', function (): void {
     $this->assertTrue($filesystem->isAbsolute('remote://file'));
 });
 
-test('test exists() method', function (): void {
+test('file exists method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('test');
     $this->assertTrue($filesystem->file($this->tempDir . '/1.txt')->exists());
@@ -131,7 +131,7 @@ test('test exists() method', function (): void {
     $this->assertTrue($filesystem->file($this->tempDir . '/2.txt')->exists());
 });
 
-test('test delete() method', function (): void {
+test('file delete method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('test');
     $this->assertTrue($filesystem->file($this->tempDir . '/1.txt')->delete());
@@ -142,41 +142,54 @@ test('test delete() method', function (): void {
     $this->assertTrue($filesystem->file($this->tempDir . '/2.txt')->delete());
 });
 
-test('test hash() method', function (): void {
+test('file hash method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('test');
     $this->assertEquals('098f6bcd4621d373cade4e832627b4f6', $filesystem->file($this->tempDir . '/1.txt')->hash());
 });
 
-test('test get() method', function (): void {
+test('file get method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('test');
     $this->assertEquals('test', $filesystem->file($this->tempDir . '/1.txt')->get());
     $this->assertEquals('test', $filesystem->file($this->tempDir . '/1.txt')->get(true));
 });
 
-test('test sharedGet() method', function (): void {
+test('file sharedGet method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/shared.txt')->put('test');
     $this->assertEquals('test', $filesystem->file($this->tempDir . '/shared.txt')->sharedGet());
 });
 
-test('test prepend() method', function (): void {
+test('file prepend method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('world');
     $this->assertEquals(11, $filesystem->file($this->tempDir . '/1.txt')->prepend('hello '));
     $this->assertEquals('hello world', $filesystem->file($this->tempDir . '/1.txt')->get());
 });
 
-test('test append() method', function (): void {
+test('file append method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('hello');
     $this->assertEquals(6, $filesystem->file($this->tempDir . '/1.txt')->append(' world'));
     $this->assertEquals('hello world', $filesystem->file($this->tempDir . '/1.txt')->get());
 });
 
+test('file isEqual method', function (): void {
+    $filesystem = new Filesystem();
+    $filesystem->file($this->tempDir . '/foo.txt')->put('Foo');
+    $filesystem->file($this->tempDir . '/foo2.txt')->put('Foo');
+    $filesystem->file($this->tempDir . '/bar.txt')->put('Bar');
+    $filesystem->file($this->tempDir . '/bar2.txt')->put('Bar2');
+    
+    $result  = $filesystem->file($this->tempDir . '/foo.txt')->isEqual($this->tempDir . '/foo2.txt');
+    $result2 = $filesystem->file($this->tempDir . '/bar.txt')->isEqual($this->tempDir . '/bar2.txt');
 
-test('test chmod() method', function (): void {
+    $this->assertEquals(true, $result);
+    $this->assertEquals(false, $result2);
+});
+
+test('file chmod method', function (): void {
     if (PHP_OS_FAMILY === 'Windows') {
         $this->markTestSkipped('The operating system is Windows');
     }
@@ -198,7 +211,7 @@ test('test chmod() method', function (): void {
     $this->assertEquals($expectedPermissions, $filePermission);
 });
 
-test('test directory chmod() method', function (): void {
+test('directory chmod method', function (): void {
     if (PHP_OS_FAMILY === 'Windows') {
         $this->markTestSkipped('The operating system is Windows');
     }
@@ -219,35 +232,35 @@ test('test directory chmod() method', function (): void {
 });
 
 
-test('test copy() method', function (): void {
+test('file copy method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('hello');
     $this->assertTrue($filesystem->file($this->tempDir . '/1.txt')->copy($this->tempDir . '/2.txt'));
 });
 
 
-test('test move() method', function (): void {
+test('file move method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('hello');
     $filesystem->file($this->tempDir . '/1.txt')->move($this->tempDir . '/2.txt');
     $this->assertTrue($filesystem->file($this->tempDir . '/2.txt')->exists());
 });
 
-test('test directory create() method', function (): void {
+test('directory create method', function (): void {
     $filesystem = new Filesystem();
     $this->assertTrue($filesystem->directory($this->tempDir . '/1')->create());
     $this->assertTrue($filesystem->directory($this->tempDir . '/1/2/3/4/')->create(0755, true));
     $this->assertTrue($filesystem->directory($this->tempDir . '/2/3/4/')->create(0755, true));
 });
 
-test('test directory ensureExists() method', function (): void {
+test('directory ensureExists method', function (): void {
     $filesystem = new Filesystem();
     $this->assertTrue($filesystem->directory($this->tempDir . '/1')->ensureExists());
     $this->assertTrue($filesystem->directory($this->tempDir . '/1/2/3/4/')->ensureExists(0755, true));
     $this->assertTrue($filesystem->directory($this->tempDir . '/2/3/4/')->ensureExists(0755, true));
 });
 
-test('test directory move() method', function (): void {
+test('directory move method', function (): void {
     @mkdir($this->tempDir . '/1');
     @mkdir($this->tempDir . '/3');
 
@@ -257,7 +270,7 @@ test('test directory move() method', function (): void {
 });
 
 
-test('test directory copy() method', function (): void {
+test('directory copy method', function (): void {
     @mkdir($this->tempDir . '/1');
     @mkdir($this->tempDir . '/3');
     @mkdir($this->tempDir . '/4');
@@ -267,18 +280,70 @@ test('test directory copy() method', function (): void {
     $this->assertTrue($filesystem->directory($this->tempDir . '/3')->copy($this->tempDir . '/4'));
 });
 
+test('directory directories method', function (): void {
+    if (PHP_OS_FAMILY === 'Windows') {
+        $this->markTestSkipped('The operating system is Windows');
+    }
 
-test('test directory delete() method', function (): void {
-    @mkdir($this->tempDir . '/1');
-    @mkdir($this->tempDir . '/1/2');
-    @mkdir($this->tempDir . '/1/2/3');
+    @mkdir($this->tempDir . '/foo');
+    @mkdir($this->tempDir . '/foo/foo-2');
+    @mkdir($this->tempDir . '/bar');
+    @mkdir($this->tempDir . '/bar/bar-2');
+    @mkdir($this->tempDir . '/zed');
+    @mkdir($this->tempDir . '/zed/zed-2');
 
     $filesystem = new Filesystem();
-    $this->assertTrue($filesystem->directory($this->tempDir . '/1')->delete());
+    $dirs = array_map(function (string $dir): string {
+        return str_replace($this->tempDir . '/', '', $dir);
+    }, $filesystem->directory($this->tempDir)->directories());
+    $this->assertTrue($dirs[0] == 'bar');
+    $this->assertTrue($dirs[1] == 'foo');
+    $this->assertTrue($dirs[2] == 'zed');
+
+    $filesystem2 = new Filesystem();
+    $dirs2 = array_map(function (string $dir): string {
+        return str_replace($this->tempDir . '/', '', $dir);
+    }, $filesystem2->directory($this->tempDir)->directories(true));
+    $this->assertTrue($dirs2[0] == 'bar');
+    $this->assertTrue($dirs2[1] == 'bar/bar-2');
+    $this->assertTrue($dirs2[2] == 'foo');
+    $this->assertTrue($dirs2[3] == 'foo/foo-2');
+    $this->assertTrue($dirs2[4] == 'zed');
+    $this->assertTrue($dirs2[5] == 'zed/zed-2');
 });
 
+test('directory files method', function (): void {
+    if (PHP_OS_FAMILY === 'Windows') {
+        $this->markTestSkipped('The operating system is Windows');
+    }
+    
+    $filesystem = new Filesystem();
 
-test('test directory clean() method', function (): void {
+    $filesystem->file($this->tempDir . '/1.txt')->put('test');
+    @mkdir($this->tempDir . '/foo');
+    $filesystem->file($this->tempDir . '/foo/1.txt')->put('test');
+    @mkdir($this->tempDir . '/foo/foo-2');
+    $filesystem->file($this->tempDir . '/foo/foo-2/1.txt')->put('test');
+    @mkdir($this->tempDir . '/bar');
+    $filesystem->file($this->tempDir . '/bar/1.txt')->put('test');
+    @mkdir($this->tempDir . '/bar/bar-2');
+    $filesystem->file($this->tempDir . '/bar/bar-2/1.txt')->put('test');
+    @mkdir($this->tempDir . '/zed');
+    $filesystem->file($this->tempDir . '/zed/1.txt')->put('test');
+    @mkdir($this->tempDir . '/zed/zed-2');
+    $filesystem->file($this->tempDir . '/zed/zed-2/1.txt')->put('test');
+
+    $filesystem = new Filesystem();
+    $this->assertTrue(count($filesystem->directory($this->tempDir)->files(true)) == 7);
+    
+    $filesystem = new Filesystem();
+    $this->assertTrue(count($filesystem->directory($this->tempDir)->files()) == 1);
+
+    $filesystem = new Filesystem();
+    $this->assertTrue(count($filesystem->directory($this->tempDir)->files(false)) == 1);
+});
+
+test('directory clean method', function (): void {
     @mkdir($this->tempDir . '/1');
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1/1.txt')->put('hello');
@@ -288,8 +353,7 @@ test('test directory clean() method', function (): void {
     $this->assertFalse($filesystem->file($this->tempDir . '/1/1.txt')->exists());
 });
 
-
-test('test glob() method', function (): void {
+test('filesystem glob method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('hello');
     $filesystem->file($this->tempDir . '/2.txt')->put('world');
@@ -302,7 +366,7 @@ test('test glob() method', function (): void {
     $this->assertEquals(0, count($glob));
 });
 
-test('test size() method', function (): void {
+test('filesystem size method', function (): void {
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
 
@@ -310,7 +374,7 @@ test('test size() method', function (): void {
 });
 
 
-test('test direcotory size() method', function (): void {
+test('direcotory size method', function (): void {
     $filesystem = new Filesystem();
     @mkdir($this->tempDir . '/1');
     @mkdir($this->tempDir . '/1/2');
@@ -322,7 +386,7 @@ test('test direcotory size() method', function (): void {
     $this->assertEquals(44, $filesystem->directory($this->tempDir . '/1')->size());
 });
 
-test('test lastModified() method', function (): void {
+test('file lastModified method', function (): void {
     $filesystem = new Filesystem();
 
     $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
@@ -331,7 +395,7 @@ test('test lastModified() method', function (): void {
     $this->assertEquals($time, $filesystem->file($this->tempDir . '/1.txt')->lastModified());
 });
 
-test('test lastAccess() method', function (): void {
+test('file lastAccess method', function (): void {
     $filesystem = new Filesystem();
 
     $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
@@ -340,7 +404,7 @@ test('test lastAccess() method', function (): void {
     $this->assertEquals($time, $filesystem->file($this->tempDir . '/1.txt')->lastAccess());
 });
 
-test('test mimeType() method', function (): void {
+test('file mimeType method', function (): void {
     $filesystem = new Filesystem();
 
     $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
@@ -348,7 +412,7 @@ test('test mimeType() method', function (): void {
     $this->assertEquals('text/plain', $filesystem->file($this->tempDir . '/1.txt')->mimeType());
 });
 
-test('test type() method', function (): void {
+test('file type method', function (): void {
     $filesystem = new Filesystem();
 
     $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
@@ -357,7 +421,7 @@ test('test type() method', function (): void {
     $this->assertEquals('dir', $filesystem->file($this->tempDir)->type());
 });
 
-test('test extension() method', function (): void {
+test('file extension method', function (): void {
     $filesystem = new Filesystem();
 
     $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
@@ -365,7 +429,26 @@ test('test extension() method', function (): void {
     $this->assertEquals('txt', $filesystem->file($this->tempDir . '/1.txt')->extension());
 });
 
-test('test basename() method', function (): void {
+test('file isEmpty method', function (): void {
+    $filesystem = new Filesystem();
+
+    $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
+    expect($filesystem->file($this->tempDir . '/1.txt')->isEmpty())->toBe(false);
+    
+    $filesystem->file($this->tempDir . '/2.txt')->put('');
+    expect($filesystem->file($this->tempDir . '/2.txt')->isEmpty())->toBe(true);
+});
+
+test('directory isEmpty method', function (): void {
+    $filesystem = new Filesystem();
+
+    $filesystem->file($this->tempDir . '/1.txt')->put('hello world');    
+    $filesystem->file($this->tempDir . '/2.txt')->put('');
+
+    expect($filesystem->directory($this->tempDir)->isEmpty())->toBe(false);
+});
+
+test('file basename method', function (): void {
     $filesystem = new Filesystem();
 
     $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
@@ -374,7 +457,7 @@ test('test basename() method', function (): void {
 });
 
 
-test('test name() method', function (): void {
+test('file name method', function (): void {
     $filesystem = new Filesystem();
 
     $filesystem->file($this->tempDir . '/1.txt')->put('hello world');
@@ -382,7 +465,7 @@ test('test name() method', function (): void {
     $this->assertEquals('1', $filesystem->file($this->tempDir . '/1.txt')->name());
 });
 
-test('test find() method', function (): void {
+test('filesystem find method', function (): void {
     @mkdir($this->tempDir . '/1');
 
     $filesystem = new Filesystem();
@@ -394,14 +477,14 @@ test('test find() method', function (): void {
     $this->assertEquals(1, count(iterator_to_array((new Filesystem)->find()->in($this->tempDir)->files()->name('*.txt'), false)));
 });
 
-test('test path() method', function (): void {
+test('directory path method', function (): void {
     $filesystem = new Filesystem();
 
     $this->assertEquals($this->tempDir . '/1.txt', $filesystem->file($this->tempDir . '/1.txt')->path());
     $this->assertEquals($this->tempDir, $filesystem->directory($this->tempDir)->path());
 });
 
-test('test macro() method', function (): void {
+test('filesystem macro() method', function (): void {
     @mkdir($this->tempDir . '/1');
     $filesystem = new Filesystem();
     $filesystem->file($this->tempDir . '/1/1.txt')->put('hello world');